+++ packages/client/src/app/ui/AuthenticatedApp/AuthenticatedApp.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import { Navigate, Route, Routes } from 'react-router-dom';
 
 import { routesConfig } from '../../lib/routes.config';
@@ -6,29 +6,19 @@ import { PAGES } from '../../lib/routes.types';
 import { GameLayout } from '../../layout/GameLayout';
 
 export const AuthenticatedApp = () => {
-  useEffect(() => {
-    const fetchServerData = async () => {
-      const url = __CLIENT_URL__;
-      const response = await fetch(url);
-      const data = await response.json();
-      console.log(data);
-    };
-
-    fetchServerData();
-  }, []);
 
   return (
-    <Routes>
-      <Route path="*" element={<Navigate to={`${PAGES.NOTFOUND}`} />} />
-      {routesConfig.map((route, idx) => (
-        <Route element={<GameLayout />} key={idx}>
-          <Route
-            path={route.path}
-            element={React.createElement(route.Component)}
-            key={idx}
-          />
-        </Route>
-      ))}
-    </Routes>
+      <Routes>
+        <Route path="*" element={<Navigate to={`${PAGES.NOTFOUND}`} />} />
+        {routesConfig.map((route, idx) => (
+          <Route element={<GameLayout />} key={idx}>
+            <Route
+              path={route.path}
+              element={React.createElement(route.Component)}
+              key={idx}
+            />
+          </Route>
+        ))}
+      </Routes>
   );
 };
