+++ packages/client/src/utils/createEmotionCache.ts
@@ -1,0 +1,5 @@
+import createCache from '@emotion/cache';
+
+export default function createEmotionCache() {
+  return createCache({ key: 'css' });
+}
