+++ packages/server/index.ts
@@ -2,6 +2,8 @@ import dotenv from 'dotenv';
 import cors from 'cors';
 import WebSocket from 'ws';
 import { createServer as createViteServer } from 'vite';
+import createCache from '@emotion/cache';
+import createEmotionServer from '@emotion/server/create-instance';
 import type { ViteDevServer } from 'vite';
 
 dotenv.config();
@@ -119,7 +121,7 @@ function start() {
     });
   }
 }
-const isDev = () => process.env.NODE_ENV === 'development';
+const isDev = process.env.NODE_ENV === 'development';
 
 async function startServer() {
   const app = express();
@@ -131,7 +133,7 @@ async function startServer() {
   const srcPath = path.dirname(require.resolve('client'));
   const ssrClientPath = require.resolve('client/ssr-dist/client.cjs');
 
-  if (isDev()) {
+  if (isDev) {
     vite = await createViteServer({
       server: { middlewareMode: true },
       root: srcPath,
@@ -145,7 +147,7 @@ async function startServer() {
     res.json('👋 Howdy from the server :)');
   });
 
-  if (!isDev()) {
+  if (!isDev) {
     app.use('/assets', express.static(path.resolve(distPath, 'assets')));
   }
 
@@ -155,7 +157,7 @@ async function startServer() {
     try {
       let template: string;
 
-      if (!isDev()) {
+      if (!isDev) {
         template = fs.readFileSync(
           path.resolve(distPath, 'index.html'),
           'utf-8'
@@ -168,22 +170,31 @@ async function startServer() {
         template = await vite!.transformIndexHtml(url, template);
       }
 
-      let render: (url: string) => Promise<string>;
+      let render: (url: string, cache: any) => Promise<string>;
 
-      if (!isDev()) {
+      if (!isDev) {
         render = (await import(ssrClientPath)).render;
       } else {
         render = (await vite!.ssrLoadModule(path.resolve(srcPath, 'ssr.tsx')))
           .render;
       }
+      const cache = createCache({ key: 'css' });
 
-      const appHtml = await render(url);
+      const { extractCriticalToChunks, constructStyleTagsFromChunks } =
+        createEmotionServer(cache);
 
-      const html = template.replace('<!--ssr-outlet-->', appHtml);
+      const appHtml = await render(url, cache);
+
+      const emotionChunks = extractCriticalToChunks(appHtml);
+      const emotionCss = constructStyleTagsFromChunks(emotionChunks);
+
+      const html = template
+        .replace('<!--ssr-outlet-->', appHtml)
+        .replace('<!--ssr-style-->', emotionCss);
 
       res.status(200).set({ 'Content-Type': 'text/html' }).end(html);
     } catch (e) {
-      if (isDev()) {
+      if (isDev) {
         vite!.ssrFixStacktrace(e as Error);
       }
       next(e);
