+++ packages/server/tsconfig.json
@@ -20,7 +20,7 @@
     "importsNotUsedAsValues": "error",
     "lib": ["DOM", "es2019", "esnext.asynciterable"],
     "types": ["node", "jest"],
-    "baseUrl": ".",
+    "baseUrl": "../",
     "paths": {
       "*": ["types/*"]
     },
